+++ analysis/tools/ValidationTool/src/test/kotlin/de/maibornwolff/codecharta/tools/validation/EveritValidatorTest.kt
@@ -10,6 +10,25 @@ class EveritValidatorTest {
     private val validator = EveritValidator(ValidationTool.SCHEMA_PATH)
 
     @Test
+    fun `should extract and validate a valid file`() {
+        validator.validate(this.javaClass.classLoader.getResourceAsStream("validCompressed.gz"))
+    }
+
+    @Test
+    fun `should throw exception if extracted file is invalid json`() {
+        assertFailsWith(JSONException::class) {
+            validator.validate(this.javaClass.classLoader.getResourceAsStream("invalidJSONCompressed.gz"))
+        }
+    }
+
+    @Test
+    fun `should throw exception on compressed json with no project`() {
+        assertFailsWith(ValidationException::class) {
+            validator.validate(this.javaClass.classLoader.getResourceAsStream("invalidProjectCompressed.gz"))
+        }
+    }
+
+    @Test
     fun `should validate valid File`() {
         validator.validate(this.javaClass.classLoader.getResourceAsStream("validFile.json"))
     }